+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`1.0.0...main`][1.0.0...main].
 ### Changed
 
 - Dropped support for PHP 7.2 ([#496]), by [@localheinz]
+- Dropped support for PHP 7.3 ([#498]), by [@localheinz]
 
 ## [`1.0.0`][1.0.0]
 
@@ -452,6 +453,7 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [#381]: https://github.com/ergebnis/phpstan-rules/pull/381
 [#395]: https://github.com/ergebnis/phpstan-rules/pull/395
 [#496]: https://github.com/ergebnis/phpstan-rules/pull/496
+[#498]: https://github.com/ergebnis/phpstan-rules/pull/498
 
 [@ergebnis]: https://github.com/ergebnis
 [@Great-Antique]: https://github.com/Great-Antique