+++ CHANGELOG.md
@@ -14,6 +14,10 @@ For a full diff see [`1.0.0...main`][1.0.0...main].
 - Dropped support for PHP 7.3 ([#498]), by [@localheinz]
 - Dropped support for PHP 7.4 ([#499]), by [@localheinz]
 
+### Removed
+
+- Removed `Expressions\NoEmptyRule` ([#525]), by [@enumag]
+
 ## [`1.0.0`][1.0.0]
 
 For a full diff see [`0.15.3...1.0.0`][0.15.3...1.0.0].
@@ -456,7 +460,9 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [#496]: https://github.com/ergebnis/phpstan-rules/pull/496
 [#498]: https://github.com/ergebnis/phpstan-rules/pull/498
 [#499]: https://github.com/ergebnis/phpstan-rules/pull/498
+[#525]: https://github.com/ergebnis/phpstan-rules/pull/525
 
+[@enumag]: https://github.com/enumag
 [@ergebnis]: https://github.com/ergebnis
 [@Great-Antique]: https://github.com/Great-Antique
 [@localheinz]: https://github.com/localheinz