+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For a full diff see [`1.0.0...main`][1.0.0...main].
 - Dropped support for PHP 7.2 ([#496]), by [@localheinz]
 - Dropped support for PHP 7.3 ([#498]), by [@localheinz]
 - Dropped support for PHP 7.4 ([#499]), by [@localheinz]
+- Added support for PHP 8.2 ([#540]), by [@localheinz]
 
 ### Removed
 
@@ -461,6 +462,7 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [#498]: https://github.com/ergebnis/phpstan-rules/pull/498
 [#499]: https://github.com/ergebnis/phpstan-rules/pull/498
 [#525]: https://github.com/ergebnis/phpstan-rules/pull/525
+[#540]: https://github.com/ergebnis/phpstan-rules/pull/540
 
 [@enumag]: https://github.com/enumag
 [@ergebnis]: https://github.com/ergebnis